requirements.txt
streamlit==1.37.0
pandas==2.2.2
langchain==0.2.12
langchain-google-genai==0.0.9
langchain-community==0.2.11
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit-authenticator==0.3.2
pyyaml==6.0.1
bcrypt==4.1.2
cohere==4.37
google-generativeai==0.3.2
python-dotenv==1.0.1
